backend/requirements.txt
django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
psycopg2-binary==2.9.11
